requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
cryptography==41.0.7
psycopg2-binary==2.9.9
requests